backend/requirements.txt
fastapi==0.121.0
uvicorn[standard]==0.38.0
motor==3.3.2
pymongo==4.6.0
python-jose[cryptography]==3.5.0
bcrypt==5.0.0
python-multipart==0.0.20
python-dotenv==1.0.0
httpx==0.28.1
websockets==15.0.1
pydantic[email]==2.5.0
pydantic-settings==2.1.0
aiofiles==25.1.0
ulid-py==1.1.0
jinja2==3.1.2
slowapi==0.1.9
python-multipart==0.0.20
cryptography==46.0.3
requests==2.31.0
